\name{fbst-package}
\alias{fbst-package}
\alias{fbst-package}
\docType{package}
\title{
\packageTitle{fbst}
}
\description{
\packageDescription{fbst}
}
\details{
Package for conducting the Full Bayesian Significance Test (FBST) for testing a sharp hypothesis against its alternative by calculating the e-value, the Bayesian evidence against the null hypothesis. The e-value is based on the relative surprise function to a reference function and the tangential set corresponding to a sharp null hypothesis. Calculation of the e-value, the p-value associated with the Bayesian e-value in favour of a sharp null hypothesis and visualisations of the FBST are provided.
\packageDESCRIPTION{fbst}
\packageIndices{fbst}

}
\author{
\packageAuthor{fbst}

Maintainer: \packageMaintainer{fbst}
}
\references{
For a details, see: https://arxiv.org/abs/2001.10577 and https://arxiv.org/pdf/2001.10577.pdf.
}

\keyword{ package }
