\name{fbst-package}
\alias{fbst-package}
\alias{fbst-package}
\docType{package}
\title{
\packageTitle{fbst}
}
\description{
\packageDescription{fbst}
}
\details{
Package for conducting the Full Bayesian Evidence Test (FBET) and the Full Bayesian Significance Test (FBST). The FBST is a Bayesian hypothesis test for testing a sharp hypothesis against its alternative by calculating the e-value, the Bayesian evidence against the null hypothesis. The FBET is a generalization of the underlying Pereira-Stern theory of the FBST and allows for testing interval hypotheses. It provides the Bayesian evidence value, or generalized e-value, which includes the e-value of the FBST as a special case. The Bayesian evidence value is based on the relative surprise function to a reference function. In the FBST, the tangential set corresponding to a sharp null hypothesis serves for quantifying the Bayesian evidence. In the FBET, the Bayesian evidence interval serves for quantifying the Bayesian evidence, which has a strong analogy to the Bayes factor. Next to the core functions, helper functions and visualizations of the results of a FBST and FBET are provided in the package.
\packageDESCRIPTION{fbst}
\packageIndices{fbst}

}
\author{
\packageAuthor{fbst}

Maintainer: \packageMaintainer{fbst}
}
\references{
For a details, see: https://arxiv.org/abs/2001.10577 and https://arxiv.org/pdf/2001.10577.pdf.
}

\keyword{ package }
